+++ src/tm-threads-common.h
@@ -108,6 +108,7 @@ typedef enum {
     TMM_UNIXMANAGER,
 
     TMM_LUALOG,
+    TMM_TLSSTORE,
     TMM_SIZE,
 } TmmId;
